fix: 3 bugs in job creation enhancement features
- fix(eval): pass dimension_name to evaluate_with_template so default
  DIMENSION_EVAL_PROMPT {dimension_name}/{job_skills} placeholders are
  properly filled; previously caused KeyError fallback with broken prompt
- fix(frontend): job-create/edit used res.data?.data instead of res.data
  for list responses (tags/dims/skills); response interceptor already
  unwraps to JSON body so .data is the array directly — tags and
  dimensions never loaded in the UI
- fix(eval): delete stale ResumeEvalDetail and ResumeSkillHit rows before
  re-evaluating an existing match to prevent duplicate records

+++ backend/app/models/resume_eval_detail.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, DECIMAL, DateTime, String, Text, TINYINT
+from sqlalchemy import Column, BigInteger, DECIMAL, DateTime, String, Text, Integer
 from sqlalchemy.sql import func
 from . import Base
 
@@ -13,6 +13,6 @@ class ResumeEvalDetail(Base):
     dimension_advantage = Column(String(500), comment="维度优点")
     dimension_disadvantage = Column(String(500), comment="维度缺点")
     ai_reasoning = Column(Text, comment="AI理由")
-    is_completed = Column(TINYINT, nullable=False, default=1, comment="是否成功完成评估：1成功，0失败")
+    is_completed = Column(Integer, nullable=False, default=1, comment="是否成功完成评估：1成功，0失败")
     error_message = Column(String(500), comment="评估失败时的错误信息")
     create_time = Column(DateTime, nullable=False, server_default=func.now(), comment="创建时间")
