+++ backend/app/models/resume_job_match.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, DECIMAL, DateTime, String, TINYINT
+from sqlalchemy import Column, BigInteger, DECIMAL, DateTime, String, Integer
 from sqlalchemy.sql import func
 from . import Base
 
@@ -13,7 +13,7 @@ class ResumeJobMatch(Base):
     final_label = Column(String(20), nullable=False, default='未达标', comment="最终标签")
     advantage_comment = Column(String(500), comment="整体优点")
     disadvantage_comment = Column(String(500), comment="整体缺点")
-    is_direct_preferred = Column(TINYINT, nullable=False, default=0, comment="是否直接优选")
+    is_direct_preferred = Column(Integer, nullable=False, default=0, comment="是否直接优选")
     error_message = Column(String(500), comment="评估失败时的错误信息")
     evaluated_at = Column(DateTime, comment="评估完成时间")
     create_time = Column(DateTime, nullable=False, server_default=func.now(), comment="创建时间")
